communicate using protobufs instead of json
this turned into a bigger project than I anticipated, mostly on the browser side. I thought that protobuf tooling for js/ts would be more mature.

I had to pull in rollup too, I was trying to avoid a bundler but protobuf has some runtime js library code that I need to bundle in and tsc won't do that.